[Bronze II] Title: 숫자, Time: 268 ms, Memory: 24008 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -5,13 +5,13 @@ class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
-        int min = Math.min(a, b);
-        int max = Math.max(a, b);
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
+        long min = Math.min(a, b);
+        long max = Math.max(a, b);
         int num = 0;
         StringBuilder sb = new StringBuilder();
-        for (int i = min + 1; i < max; i++) {
+        for (long i = min + 1; i < max; i++) {
             sb.append(i).append(" ");
             num++;
         }
